requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
fastapi[all]==0.110.2
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.3
    # via fastapi
mako==1.3.3
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
orjson==3.10.1
    # via fastapi
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.7.1
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.18.2
    # via pydantic
pydantic-extra-types==2.7.0
    # via fastapi
pydantic-settings==2.2.1
    # via fastapi
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   fastapi
    #   uvicorn
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.16
    # via -r requirements.in
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
ujson==5.9.0
    # via fastapi
uvicorn[standard]==0.29.0
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
